Created Tools folder
This folder will contain all tools that are useful to support installation and testing of BashUtils.
This mean all content that should not be used by a BashUtils user.
Moved the TESTS folder from root to this folder.
Updated all Test scripts so that they point to new location of test common files.